pyproject.toml
[tool.poetry]
name = "dicebot"
version = "0.1.0"
description = ""
authors = ["Fabian Schlager <fabian.schlager@skyr.at>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.2.0"
pre-commit-hooks = "^2.5.0"
neovim = "^0.3.1"
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
